Add logic for fifth LED hole
More LEDs is always better.  Always!

Also, at .6 watts each, this is only 3 watts total @ 5 volts for the five LEDs.  That's 0.6 amps for the LEDs all at full brightness white, plus barely 150mA for the Pi Zero W.  We're under an amp even running a Pi 3 on this, which is well within the capability of any common USB adapter, and not going to overload the circuit traces.  Just need to be a bit careful about the sound output, as _that_, at high volume with a high-power speaker, would require a bigger supply.  But to just play back a short wav file or whatever, we'll still be just fine - especially with the tiny speaker I intend to use.